More bit vector fixes

+++ include/coreir/ir/dynamic_bit_vector.h
@@ -1086,7 +1086,7 @@ namespace bsim {
   quad_value_bit_vector
   zero_extend(const int outWidth, const quad_value_bit_vector& in) {
     quad_value_bit_vector res(outWidth, 0);
-    for (uint i = 0; i < in.bitLength(); i++) {
+    for (uint i = 0; i < (uint) in.bitLength(); i++) {
       res.set(i, in.get(i));
     }
